pyproject.toml
[project]
name = "mcp-build"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-mcp-adapters>=0.1.1",
    "neovim>=0.3.1",
]
